one dropdown box works, need more fixing
