+++ src/components/SearchFilter/Search_help.js
@@ -26,14 +26,75 @@ const SearchHelp=()=>{
          getUsers();
          searchTerm.current.focus();
    },[])
+
+////////////////
+   const [filterData,setFilterData]=useState({
+      location:'',
+      category:'',
+      period:'',
+    })
+  
+    console.log(filterData)
+  
+    const [dataAlreadyFiltered,setDataAlreadyFiltered]=useState([data]);
+  
+    const updateFilterData=(e)=>{
+       const {name,value}=e.target;
+       setFilterData(preVal=>{
+          return {
+            ...preVal,
+            [name]:value,
+          }
+       })
+    }
+    
+    
+    
+    useEffect(()=>{
+
+      const updateResults=()=>{
+      
+         if(filterData.location!=='' ||
+            filterData.category!=='' ||
+            filterData.period!==''
+         ){
+            const test=data.filter(el=>
+
+               
+            ( (el.location===filterData.location || el.category===filterData.category || el.period===filterData.period) || (el.location===filterData.location && el.category===filterData.category)
+            ))        
+            
+            
+            return setDataAlreadyFiltered(test);
+           
+            
+         }else{ return  setDataAlreadyFiltered(data)
+          
+          
+         }
+        
+       
+      }
+
+      console.log(filterData);
+      updateResults();
+    },[filterData, data])
+
+    /////////////
+
+
+    console.log(dataAlreadyFiltered);
+
+
+
    return (
       <div>
-         <Search update={updateSearch} word={searchTerm} data={data}/>
+         <Search update={updateSearch} word={searchTerm} filterData={filterData} updateFilterData={updateFilterData} updatedata={data}/>
          
          {data!==[] && (
                   <Help
                      filter={searchValue}
-                     data={data}
+                     data={dataAlreadyFiltered}
                   />
             )
          }
